完善mysql

+++ evaluation.sh
@@ -9,28 +9,35 @@ SRC_DIR=/data/cbj/temp
 cd ~/destor
 remake
 
-resetAll
-rm -r ~/destor/log/time*
-mkdir -p ~/destor/log/time
-
 [[ "$*" =~ "-chk" ]] && chk=1 || chk=0
 [[ "$*" =~ "-bkp" ]] && bkp=1 || bkp=0
 
-mysql -uroot -proot -e "
-  use CBJ;
-  drop table if exists kvstore;
-  create table kvstore ( k BINARY(32) PRIMARY KEY, v BINARY(8));
-"
-
 set -x
-# basic test
-mkdir -p ${DST_DIR}${RESTORE_ID}
-destor ${SRC_DIR} "${CONFIG}" > ${LOG_DIR}/${RESTORE_ID}.log
-cp -r ${WORKING_DIR} ${WORKING_DIR}_bak
+rm -rf ~/destor/log/time*
+if [[ $bkp == 1 ]]; then
 
-destor -r0 ${DST_DIR}${RESTORE_ID} "${CONFIG}"
-mv ~/destor/log/time ~/destor/log/time${RESTORE_ID}
-mkdir ~/destor/log/time
+    resetAll
+        
+    mysql -uroot -proot -e "
+    use CBJ;
+    drop table if exists kvstore;
+    create table kvstore ( k BINARY(32) PRIMARY KEY, v BINARY(8));
+    "
+
+    set -x
+    # basic test
+    mkdir ~/destor/log/time
+    mkdir -p ${DST_DIR}${RESTORE_ID}
+    destor ${SRC_DIR} "${CONFIG}" > ${LOG_DIR}/${RESTORE_ID}.log
+    cp -r ${WORKING_DIR} ${WORKING_DIR}_bak
+
+    if [ $chk -eq 1 ]; then
+        destor -r0 ${DST_DIR}${RESTORE_ID} "${CONFIG}"
+    fi
+    mv ~/destor/log/time ~/destor/log/time${RESTORE_ID}
+
+fi
+
 let ++RESTORE_ID
 
 function update() {
@@ -38,12 +45,15 @@ function update() {
     rm -r ${WORKING_DIR}
     cp -r ${WORKING_DIR}_bak ${WORKING_DIR}
     
+    mkdir -p ~/destor/log/time
     mkdir -p ${DST_DIR}${RESTORE_ID}
     destor -u0 ${SRC_DIR} -i"$1" "${CONFIG}" > ${LOG_DIR}/${RESTORE_ID}.log
-    rm ${WORKING_DIR}/container.pool
-    destor -n1 ${DST_DIR}${RESTORE_ID} "${CONFIG}"
+
+    if [ $chk -eq 1 ]; then
+        rm ${WORKING_DIR}/container.pool
+        destor -n1 ${DST_DIR}${RESTORE_ID} "${CONFIG}"
+    fi
     mv ~/destor/log/time ~/destor/log/time${RESTORE_ID}
-    mkdir ~/destor/log/time
 
     let ++RESTORE_ID
 }
@@ -64,7 +74,9 @@ mysql -uroot -proot -e "
 "
 update 2
 
-compareRestore
+if [ $chk -eq 1 ]; then
+    compareRestore
+fi
 
 if [ ${flag} -eq 0 ]; then
     make clean -s
